An addition to the collection
This is another code for the program of "swapping two numbers" as it is without using a temporary variable.
It is short and simple but has its own limitations of memory buffer due to bit-wise operators and multiplication or division, if a nos. are very big to swap, the contiguous memory blocks might not be able to store the size. But that is obviously a rare case.
Thank you!